[TDO-108] Add option to provide custom request adapter (for use with k6)
Add the option to set a custom request adapter (an axios feature). Adapters should return a promise and supply a valid response.

For load testing, we want to construct requests but not send them (as requests are made through k6). With these changes, setting the adapter to `() => Promise.resolve({})` can achieve this.