Fix the default access denied handler in `SwsClient`
Previously it would attempt to retry the request with an access token refresh if the access token was empty at the time of the error, however we need to also check if the refresh
token is not empty, otherwise it will be a wasted call. This commit also makes the default access denied handler invoke the default error handler if the conditions for a token
refresh are not satisfied.